Add ability to log job output with details
This allows the user to specify `--log-dir DIR` which will cause a log
file to be created in `DIR`. The log file will be named with the job
start date, time, executable file name, and child PID.

The log file will contain the full command line and the start time at
the top followed by a blank line then a record of the events received,
e.g. output on stdout or stderr, along with the relative time since the
process started.

Example:

```text
Command: ["my-script.sh", "arg1", "arg2"]
Start: 2023-05-15T18:31:41.509380000-07:00

+0.000 out foo bar hey
           next line whatever
+0.010 err NOTICE bar hey
+0.100 out mixed
+0.200 \ out .
+0.340 \ out .
+0.610 \ out  done
+0.610 exit 0
```

This also updates the MSRV to 1.65 to support time crate.