Add `Event::exit_code()`
Add convenience function to get the UNIX-style exit code from a child
process. Using `status.code()` doesn’t return a code if the child was
killed by a signal; this returns 128 + the signal number in that case.

This also renames the `command` module tests for clarity and adds checks
to confirm that `exit_code()` works properly.