Retry generating log file name up to 100 times
The default naming scheme for log files includes the time down to the
second (including the timezone) and the process ID, which should almost
always be unique. This adds a feature to append a number and retry
creating the file if it already exists. It will retry up to 100 times.