Clarify that reading 0 bytes doesn’t crash
`Child::read()` always set the state to `Polling` after reading fewer
than the maximum bytes. This switches to also setting the state to
`Polling` in `Child::next_event()` when `read()` returns 0 bytes so that
it is obvious that it moves on to `poll()` and doesn’t reach the
`unreachable!()` at the bottom of the function.