Expand linting
This enables `clippy::nursery` and a bunch of lints from
`clippy::restriction`, and fixes all violations. It adjusts lints to
match what I’m working on in other projects.

This is probably not the final lint list — there are comments on the
lint list reflecting that I’m unsure about some of the lints.