Honor timeouts after child has closed stdout/err
Previously we used the standard `wait()` function once the output
streams were closed to wait for the child to exit. Unfortunately, this
does not allow for a time out.

This is particularly a problem if the child process closes `stdout` and
`stderr` then does more work.

This adds a custom `wait()` method to our `Child` struct that honors
timeouts by calling `try_wait()` in a busy loop. It might be better to
use an alarm signal to wake process within `wait()`, but that could
cause problems with other code that uses signals, or if the process has
multiple threads.

Fixes: #44 — If child process closes stdout/stderr the timeout will no longer work