`job_logger::parser` and `replay` example
This adds a `replay` example that parses structured log files and
reproduces the output as if it was just run through `cron-wrapper`. It
doesn’t presently parse errors so it prints their `Debug` output.

This is just a starting point, but I think it’s solid enough to merge.

There are no integration tests, and most of the parser is missing unit
tests.

This should probably be integrates into the cron-wrapper binary as a
subcommand. Possibly it should be available as a symlink, i.e. the
behavior should change based on `argv[0]`.

Fixes #32: Add replay command to replay log files