Escape structured log values and missing newlines
This makes structured log output escape most control characters other
than tab and newline. It also escapes backslashes.

This also changes the way output without a trailing newline is handled.
Previously, we indicated if a line was a “continuation” by prefixing it
with a backslash and a space. This switches to indicating that a line
didn’t have a trailing newline by putting a backslash at the end of the
line, much as we might do in a Rust string literal.