Don’t panic parsing time offsets over `Duration::MAX`
Previously the parser would panic if it encountered a time offset with
more than `u64::MAX` seconds. This switches to returning a slightly
nicer parse error.

This also switches a couple of `expect()`s that are guarded against in
the same function to `unwrap()`. It’s just easier to read.