`Timeout::elapsed_rounded_to()`
Previously, `Timeout::elapsed_rounded()` was hard coded to round to the
nearest millisecond. This renames the function and adds a parameter to
specify what factor the duration should be rounded to.

For example, `timeout.elapsed_rounded_to(Duration::from_millis(1))`
returns the elapsed time rounded to the nearest millisecond.

This also adds a private function to round `Duration`s more generally,
and a private function to create a `Duration` from `u128` of
nanoseconds (`Duration::from_nanos()` takes a `u64`).