Add `Timeout::check_expired_within()`
`check_expired_within()` allows for checking timeouts when the
resolution is larger than 1 nanosecond.

For example, [`poll.2`] takes a timeout measured in milliseconds. If we
didn’t take the resolution into account, it we might try to re-run
`poll()` despite the timeout having already elapsed from its
perspective.

This also changes `Timeout::check_expired()` to operate on a resolution
of nanoseconds (the smallest unit of time we can represent with
`Duration`) instead of working on milliseconds. To get the previous
behavior we have to change `timeout.check_expired()` to
`timeout.check_expired_within(Duration::from_millis(1))`.

[`poll.2`]: https://man7.org/linux/man-pages/man2/poll.2.html