Use `lints.toml` instead of `[lints]` table in `Cargo.toml`
This reverts commit 1bfba8b6ffe291d3e1b52923cd74f8f657e86c2e and adds a
few other changes.

It appears that the `[lints]` table doesn’t apply rules in the order
they’re listed in `Cargo.toml`, but rather in alphabetical or some other
order. This as the affect that some lints cannot be turned off if the
group they belong to is turned on.

For example, if `clippy::nursery` is set to `warn`, then it is
impossible to set `clippy::option_if_let_else` to `allow`.