Add `paused` parameter to `PausableWriter::stdout()`
Previously, if you wanted to start `PausableWriter` unpaused you had to
call a function that could return an error. In practice the buffer would
be empty so the IO code didn’t actually need to be called. This adds a
`paused` parameter to the constructor to avoid the whole problem.